CSExporter/requirements.txt
crowdstrike-falconpy>=1.4.5
pandas>=2.1
openpyxl>=3.1
python-dotenv>=1.0
pyreadline3